+++ src/classes/modules/java.base/java/lang/System.java
@@ -23,8 +23,8 @@ import java.io.PrintStream;
 import java.nio.channels.Channel;
 import java.util.Map;
 import java.util.Properties;
-import jdk.internal.misc.JavaLangAccess;
-import jdk.internal.misc.SharedSecrets;
+import jdk.internal.access.JavaLangAccess;
+import jdk.internal.access.SharedSecrets;
 import jdk.internal.reflect.ConstantPool;
 import sun.nio.ch.Interruptible;
 import sun.reflect.annotation.AnnotationType;
