+++ src/classes/modules/java.base/jdk/internal/access/JavaIODeleteOnExitAccess.java
@@ -16,38 +16,17 @@
  * limitations under the License.
  */
 
-package jdk.internal.misc;
-
-import jdk.internal.reflect.ConstantPool;
-import sun.nio.ch.Interruptible;
-import sun.reflect.annotation.AnnotationType;
+package jdk.internal.access;
 
 /**
  * this is a placeholder for a Java 6 class, which we only have here to
  * support both Java 1.5 and 6 with the same set of env/ classes
  *
- * see sun.misc.SharedSecrets for details
+ * see sun.msic.SharedSecrets for details
  *
  * <2do> THIS IS GOING AWAY AS SOON AS WE OFFICIALLY SWITCH TO JAVA 6
  */
-
-public interface JavaLangAccess {
-
-    ConstantPool getConstantPool(Class<?> klass);
-
-    void setAnnotationType(Class<?> klass, AnnotationType annotationType);
-
-    AnnotationType getAnnotationType(Class<?> klass);
-
-    <E extends Enum<E>> E[] getEnumConstantsShared(Class<E> klass);
-
-    void blockedOn(Thread t, Interruptible b);
-
-    void registerShutdownHook(int slot, Runnable r);
-
-    int getStackTraceDepth(Throwable t);
-
-    StackTraceElement getStackTraceElement(Throwable t, int i);
-
-    Module defineUnnamedModule(ClassLoader loader);
+public interface JavaIODeleteOnExitAccess extends Runnable {
+  @Override
+  public void run();
 }
