+++ src/classes/modules/java.base/jdk/internal/access/JavaLangAccess.java
@@ -16,17 +16,38 @@
  * limitations under the License.
  */
 
-package jdk.internal.misc;
+package jdk.internal.access;
+
+import jdk.internal.reflect.ConstantPool;
+import sun.nio.ch.Interruptible;
+import sun.reflect.annotation.AnnotationType;
 
 /**
  * this is a placeholder for a Java 6 class, which we only have here to
  * support both Java 1.5 and 6 with the same set of env/ classes
  *
- * see sun.msic.SharedSecrets for details
+ * see sun.misc.SharedSecrets for details
  *
  * <2do> THIS IS GOING AWAY AS SOON AS WE OFFICIALLY SWITCH TO JAVA 6
  */
-public interface JavaIODeleteOnExitAccess extends Runnable {
-  @Override
-  public void run();
+
+public interface JavaLangAccess {
+
+    ConstantPool getConstantPool(Class<?> klass);
+
+    void setAnnotationType(Class<?> klass, AnnotationType annotationType);
+
+    AnnotationType getAnnotationType(Class<?> klass);
+
+    <E extends Enum<E>> E[] getEnumConstantsShared(Class<E> klass);
+
+    void blockedOn(Thread t, Interruptible b);
+
+    void registerShutdownHook(int slot, Runnable r);
+
+    int getStackTraceDepth(Throwable t);
+
+    StackTraceElement getStackTraceElement(Throwable t, int i);
+
+    Module defineUnnamedModule(ClassLoader loader);
 }
