+++ src/classes/modules/java.base/jdk/internal/access/JavaNioAccess.java
@@ -15,20 +15,19 @@
  * See the License for the specific language governing permissions and 
  * limitations under the License.
  */
-
-package jdk.internal.misc;
-
-import java.net.URLClassLoader;
-import jdk.internal.loader.URLClassPath;
+package jdk.internal.access;
 
 /**
- * this is a placeholder for a Java 6 class, which we only have here to
- * support both Java 1.5 and 6 with the same set of env/ classes
- *
- * see sun.msic.SharedSecrets for details
- *
- * <2do> THIS IS GOING AWAY AS SOON AS WE OFFICIALLY SWITCH TO JAVA 6
+ * <2do> this is not supported yet
  */
-public interface JavaNetAccess {
-    URLClassPath getURLClassPath (URLClassLoader ucl);
+public interface JavaNioAccess {
+  
+  interface BufferPool {
+    String getName();
+    long getCount();
+    long getTotalCapacity();
+    long getMemoryUsed();
+  }
+
+  BufferPool getDirectBufferPool();
 }
